Linkedin Content Creator using LLM/requirements.txt
streamlit==1.35.0
langchain==0.2.14
langchain-core==0.2.39
langchain-community==0.2.12
langchain_groq==0.1.9
pandas==2.0.2
python-dotenv==1.0.1
plotly==5.21.0
stability-sdk==0.8.6
Pillow==10.4.0
matplotlib==3.8.4
seaborn==0.13.2
